Python_Dictionary
I have created and interactive dictionary using python. I used a .json file which has a dictionary consisting of huge collection of words. 
The program takes input from the user and checks for the word in the json file. If it finds a match it returns it, otherwise it suggests words that closely match the input word. If none of them match, it prints that the word doesn't exist.